requirements.txt
# Agent framework (with OpenAI)
pydantic-ai[openai]>=0.0.14

# Sandbox API client
httpx>=0.27.0

# Vector store (local, lightweight)
chromadb>=0.4.22

# Structured memory
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Sandbox server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Config
python-dotenv>=1.0.0

# Docker SDK for sandbox container management
docker>=7.0.0

# Retries for sandbox API client
tenacity>=8.0.0

# Observability (optional)
logfire>=0.1.0

# Embeddings for vector store
sentence-transformers>=2.2.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
